Migrate faceForward tests
